requirements.txt
streamlit
genrativeai
PyMuPDF
nest_asyncio
googletrans==4.0.0rc1
gtts
python-dotenv